Add custom soft errors for dep only incompatible
Summary:
See D70855390 for full context.

This diff adds the full behavior of custom soft errors. When buckconfig `buck2.dep_only_incompatible_info` is set, buck2 will check if the configured target node that fails the check belongs in any of the target patterns. If it does, it will fire a custom soft error for the corresponding target pattern.

Reviewed By: IanChilds

Differential Revision: D70863495

fbshipit-source-id: 4f53fc4524884c966c93404154e3e9dcf249253d