implement opt-by-default configuration transition
Summary:
This diff adds a buck2 cfg transition into the python_targets.
Transitions are used to change configurations when entering or exiting build nodes. In this case we add a transition to python_binary and pyton_unittest targets which will apply the opt-by-default constraints to the targets (and transistive deps)

The key point of this implementation is that constraints will be applied regardless of the top level target, furthermore applied constraints will be limited to python_* nodes and their dependencies, ie won't leak to cpp_binaries that depend on python_binaries

For now this should be a no-op, moving forward for the default-on will be implemented by applying the -enabled constraint to every target in fbcode and the conditional modifiers will be removed.

#buildmore

Reviewed By: zsol

Differential Revision: D70286674

fbshipit-source-id: bf83e25850a669ff344c5c65e88dc0e49f30fb4b