Validate the soft error categories when constructing DepOnlyIncompatibleInfo provider
Summary: Forgot to do this in the diff that created DepOnlyIncompatibleInfo provider. We need to validate that all the categories provided are valid logview categories when creating the provider. Otherwise, we will throw an error only when we attempt to log the quiet soft error if we find an invalid category - which would be quiet bad.

Reviewed By: jdonald

Differential Revision: D71423540

fbshipit-source-id: 8c98cffa501afe9f6c0c1641c15908aed4639167