Implement A/B testing components in Kotlin Toolchain
Summary:
### Context
We are exploring the integration of the Kotlin Incremental Compiler into Buck for the incremental compilation of Kotlin source files. The goal is to improve build speed by only recompiling modified source files within a target, rather than rebuilding the entire target as is currently done.

To ensure a smooth rollout, we would like to experiment with an experimentation framework for KSIC. The primary goal of this framework is to detect any unexpected behavior or error increases.

### This diff
Implements A/B testing components for Kotlin toolchain. The main reason for adding extra target was to avoid adding dependencies on JSON library into prelude//toolchains/android/src/com/facebook/buck/jvm/kotlin:steps target.

----
Internal
[Doc](https://docs.google.com/document/d/1-zhGPkmt3I2os-kcIaFk9SvT6odt2AJWswSY-J_6AWs/edit?usp=sharing)|[Experiment doc](https://docs.google.com/document/d/1BD1uRf7B3iwBm9sJ6EIabqMFh-sHR6Lw9UefNd6Sw_4/edit?usp=sharing)

Example of usage: D71405129

Reviewed By: mzlee

Differential Revision: D71405130

fbshipit-source-id: a94fdeaa3d16913fb23dea216cf8fe6a43de4c30