Erlang elp.bxl: make apps actually overwrite prelude
Summary:
The `elp.bxl` query for the project model in ELP does not load prelude applications unless they are in an application dependency chain somewhere.

But sometimes we are explictly listing a prelude target in the set of ones we want to process.

To deal with this, we first populate our apps list with an entry tagged prelude, which can then be overwritten by one tagged app, if this is the case.

This did not work before, because we used a list, not a set.

This diff switched to using a set, indexed by the target label.

Reviewed By: Balajiganapathi

Differential Revision: D71320108

fbshipit-source-id: 04c7f9e62dc3dace883cd0d4a7c5be7f22139834