Fix setting of supported_build_modes
Summary:
While the default of supported build modes is APPLE_BUILD_MODES:
https://www.internalfb.com/code/fbsource/[efcfb52a1dc08d5ca9f9756dcb3719dc455e98fe]/fbcode/buck2/prelude/platforms/apple/platforms.bzl?lines=135%2C155

We pass it an empty list:
https://www.internalfb.com/code/fbsource/[ad8a16ff5bcb0d1f458bc84871d294fefeb1f20d]/fbcode/buck2/prelude/apple/apple_native.bzl?lines=67-69%2C72%2C79

This regressed in D70800926, breaking project generation:
```
error: Object of type `NoneType` has no attribute `raw_target`
  --> xplat/buck2/xcode_project/platforms.bxl:14:33
   |
14 |     target_platforms = {config: platforms.get(config).raw_target() for config in build_configurations.keys()}
   |                                 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
   |
```

This is a quick fix to unbreak things. We'll look for a better approach and adding a quicker test in a followup diff.

Reviewed By: ChrisRicca

Differential Revision: D71574440

fbshipit-source-id: 8d04bbe74c06977004c030d34b154c5c6f1025ba