Move LoggerCatKotlinCDScribeLogger into a separate target
Summary:
### Context
The Scribe logger was integrated into the Android Build Toolchain to collect data from KotlinCD steps. However, with the recent transition of ABT to open-source, we can no longer include Meta-specific code in the OSS release, as it would cause build failures.

### This Stack
This stack aims to use ServiceLoader to inject the correct implementation based on the repository source (OSS or Meta).

**Final structure**

 {F1976250827}

### This diff
Moves LoggerCatKotlinCDScribeLogger into a separate target for a better dependency management at later diffs.

Reviewed By: hick209

Differential Revision: D71486549

fbshipit-source-id: d746059cdfd70977a312c55f58e76acc0d35f343