Rename :kotlincd_scribe_logger target to :kotlincd_logger
Summary:
### Context
The Scribe logger was integrated into the Android Build Toolchain to collect data from KotlinCD steps. However, with the recent transition of ABT to open-source, we can no longer include Meta-specific code in the OSS release, as it would cause build failures.

### This Stack
This stack aims to use ServiceLoader to inject the correct implementation based on the repository source (OSS or Meta).

**Final structure**

 {F1976250827}

### This diff
Renames the target to match naming of its classes.

Differential Revision: D71486545

fbshipit-source-id: f5c8bcf40dd5352f38e8108124801539fbe28fbc