Add tags attribute
Summary:
`#[derive(buck2_error_derive::Error)]` supports passing an expression (typically a function) to `tag` to determine error tags, but it's not possible to return an optional tag or variable number of tags.
This adds a `tags` field that is an expression that returns `Vec<ErrorTag>` so that some tags can be optional.

Currently we convert error types to `buck2_error::Error` to do this but there are places where it's not ideal to lose the underlying error type.

Reviewed By: iguridi

Differential Revision: D71491703

fbshipit-source-id: 5b94a14f797196a38272715c92d9cb816a3c4502