Create :noop target in ABT
Summary:
### Context
The Scribe logger was integrated into the Android Build Toolchain to collect data from KotlinCD steps. However, with the recent transition of ABT to open-source, we can no longer include Meta-specific code in the OSS release, as it would cause build failures.

### This Stack
This stack aims to use ServiceLoader to inject the correct implementation based on the repository source (OSS or Meta).

**Final structure**

 {F1976250827}

### This diff
Creates KotlinCDNoopLogger in the :noop target to be used as an implementation when building for OSS.

Reviewed By: jselbo

Differential Revision: D71559748

fbshipit-source-id: d6c085130afc74d22ac5f52dc7c6204fc3025c19