Move QE2 out of prelude
Summary:
### Context
The QE2 framework was integrated into the Android Build Toolchain to conduct an experiment for KSIC. However, with the recent transition of ABT to open-source, we can no longer include Meta-specific code in the OSS release, as it would cause build failures.

### This Stack
This stack implements ServiceLoader-driven injection to select the correct implementation based on the repository source (OSS or Meta), aligning with the existing Scribe implementation pattern.

### This diff
Moves QE2 target into fbsource.

Reviewed By: RuslanLa

Differential Revision: D71625526

fbshipit-source-id: 31de0e9bc3232a464f22788cf54ca0be458409d2