Move QE2 components into qe2 target
Summary:
### Context
The QE2 framework was integrated into the Android Build Toolchain to conduct an experiment for KSIC. However, with the recent transition of ABT to open-source, we can no longer include Meta-specific code in the OSS release, as it would cause build failures.

### This Stack
This stack implements ServiceLoader-driven injection to select the correct implementation based on the repository source (OSS or Meta), aligning with the existing Scribe implementation pattern.

### This diff
Moves QE2 components into a separate target for easier dependency management.

Reviewed By: RuslanLa

Differential Revision: D71625525

fbshipit-source-id: 96a2406fa298d6f15bf75d08cf9b592ede184b05