set debug app in buck install
Summary:
We currently only set debug app when we run installation with `buck install -r`. Setting debug app is important since it prevents the app from ANRs.
We could optimize this by calling `adb shell am set-debug-app ` when people attacher debugger. However, calling that will stop the running app which will be a disruptive experience.
The idea here is to simply set the debug app when we do `buck install` even if `-r` is not specified.

Reviewed By: nghuiqin, hick209, jselbo

Differential Revision: D71365023

fbshipit-source-id: 32fb56fbe4cec922032987b071a8479302b96f3f