Add test for error handler fail on None input
Summary:
There are scenarios when the toolchain doesn't contain an error handle and would therefore pass in a `None`. This issue has been observed in both cxx D61277187 and python D71658974. The issue can be bypassed but it would make sense to allow error handler to take in `None` and just not do anything rather than failing to make the experience better.

Adding a test to repro this behavior to be fixed in the next diff

Reviewed By: JakobDegen

Differential Revision: D71767826

fbshipit-source-id: a36e2be7b6e14df375b166e232f079046d763898