Improve default for supported_build_modes
Summary:
The default for `supported_build_modes` is `APPLE_BUILD_MODES`.

However, when passing arguments, we need to handle the case where we don't override the default.

In D71574440, we did this passing optional keyword arguments, to work around setting the default to `[]` or `None`.

In this diff, we clean this up to handle the logic better within the function itself. Now, users of the API can either pass a value, or nothing at all, in which case, we'll utilize the right default.

Reviewed By: IanChilds

Differential Revision: D71780671

fbshipit-source-id: 687d4bd992c4ac875fa71fd1e06cc55dcdf83cf4