Handle CommandResult Error and add to invocation record
Summary:
More context in the previous diff description.

Invocation Recorder only record errors in `CommandEnd` but not `CommandResults`, even though `CommandResults` has a more complete set of errors. This is likely the case because of `build` commands. `build` works differently, not all errors from `build` are recorded in `CommandResults` which we should fix and then rely on `CommandResults` only

Reviewed By: christolliday

Differential Revision: D71657593

fbshipit-source-id: 36490fb39479cd983fc3afd0fbddc1e15291be3c