Store output/error in memory instead of files
Summary:
Store output/error in memory instead of files.

1. Outputs and errors in bxl most cases will not have too many and users could write the output into files if the output is too large. We are safe to make such changes
2. It removes the write to file side effect for computing bxl dice key
3. It will also solve such bug in bxl:
https://fb.workplace.com/groups/buck2users/permalink/3903296533260032/
In the example from the post, bxl key depends on a action key. In the second run, it fails and the output file path will be override. In the third, it changes file back to the first run, it will use the first run's dice cache, and the output will be empty.

Reviewed By: scottcao

Differential Revision: D71357556

fbshipit-source-id: 00b223c31c08b321e02f426a8a7998fab8f155c8