Framework to add action_inputs_hash logging to actions
Summary:
We are still looking into the best way to do action deduplication. One option is to make a hash of the "inputs" to an action.

This stack adds some logging of an `action_inputs_hash`. When an action is registered, it can supply such a hash and that hash will be stored on the action for logging, and it will also be given to any output artifact produced by the action (by passing it to the output artifact at `bind` time).

Reviewed By: JakobDegen

Differential Revision: D70894614

fbshipit-source-id: f7d75a303b5c340612cfe752e8e111e1ad1a32e0