Add buck config to enable computing an action inputs hash at analysis time
Summary:
I want to log an action inputs hash so that we can see if this is a sensible approach for action deduplication.

However, I want to do it probably just for a short period of time to gather data, and then turn it off again. So, here is a buckconfig to allow us to do that.

Reviewed By: JakobDegen

Differential Revision: D71044707

fbshipit-source-id: af38a2efc11ce2f92098b20c89a97013a506f825