Make UnregisteredAction use less memory
Summary:
Probably not the biggest deal, but when I look at the output of `buck2 debug allocative`, I see a bunch of optional fields on `UnregisteredRunAction` using a surprising amount of memory.

{F1976294251}

IIUC, that is because when we use an `Option<RemoteExecutorCustomImage>` on the `UnregisteredRunAction`, it's still allocating the amount of space required for `RemoteExecutorCustomImage`, even though for most builds it's just `None`.

Please, "roast my Rust". I'm not sure if wrapping this with a `Box` (to make the space allocated for the field in `UnregisteredRunAction` just be a single pointer) is the best solution here, is there something better we can do? What's the usual pattern for something like this? Thanks!

Reviewed By: blackm00n

Differential Revision: D71620933

fbshipit-source-id: c5c48a3ca5d25c724bd78f4cce66eca773f44328