Propagate string_tags to ActionError
Summary:
Action errors are missing the TCodeReasonGroup returned from RE, which is useful for categorization, because they are string tags, unlike TCode tags.

This could be less error prone but we only use `provide_metadata` in one place, which would ideally be removed but that requires a larger cleanup to error formatting.

Reviewed By: JakobDegen

Differential Revision: D71853122

fbshipit-source-id: 16c75f5e32d317aeec4dac9437a9fc6de3960822