Recategorize RE_FAILED_PRECONDITION as an input error
Summary:
We're seeing frequent errors tagged RE_FAILED_PRECONDITION with the message "Platform {platform} not listed for use case {use case}".

After the previous diff these should also have the TCodeReasonGroup to USER_BAD_REQUEST. We should probably tag all RE errors with `USER_` as input errors, but we also need to not override the category from the TCode. It seems likely that every error with this TCode should be an input error.

Reviewed By: JakobDegen

Differential Revision: D71855648

fbshipit-source-id: d45aad5fbf4effbf79575ca8efc72f3f9eca006e