Parse sub-commands, inputs, and outputs from nvcc dryrun dump
Summary:
This diff implements the functionality in wrap_nvcc, where if we're running nvcc with -dryrun, then generate a dependency graph of the nvcc sub-commands. Some examples of the dag can be found in the test fixtures. The dag is serialized as a json file that consists of a list of graph nodes, where each node contains:
* The category name which will be used to name the Buck action.
* The command tokens.
* List of input files.
* List of output files.
* List of hidden dependencies (dependencies that don't show up in the command).

The input, output, and hidden files will be bound to Buck Artifacts in a later diff. Because we need to replace the command line arguments with the actual Buck Artifacts, The files from the command are picked out and replaced with `{input}` and `{output}` placeholders. These tokens will be replaced with Artifacts later in Buck prelude.

Add unit tests.

Reviewed By: danzimm

Differential Revision: D71352313

fbshipit-source-id: ba758f627b2fd6d6eb74353b13188f8f4bf53238