Introduce QE2 enablement detection in rule layer
Summary:
### Context
We are exploring the integration of the Kotlin Incremental Compiler into Buck for the incremental compilation of Kotlin source files. The goal is to improve build speed by only recompiling modified source files within a target, rather than rebuilding the entire target as is currently done.

To ensure a smooth rollout, we would like to experiment with an experimentation framework for KSIC. The primary goal of this framework is to detect any unexpected behavior or error increases.

As the first step, I will focus on targeting incremental pure Kotlin targets that do not use additional plugins or annotation processors. Since incremental compilation is enabled on mixed modules and modules using additional plugins or annotation processors, we cannot rely solely on incremental flag as an indicator. Therefore, I have introduced an additional logic that evaluates whether experimentation is enabled for a given target.

### This diff
Addresses this [comment](https://www.internalfb.com/diff/D71822715?dst_version_fbid=647706011342825&transaction_fbid=1906302516864730) and introduces automatic qe2 enablement detection in the rule layer.
 ---
Internal
[KSIC Experiment Doc](https://docs.google.com/document/d/1BD1uRf7B3iwBm9sJ6EIabqMFh-sHR6Lw9UefNd6Sw_4/edit?usp=sharing)

Differential Revision: D71890005

fbshipit-source-id: b0c503fd789bc5cfbb4ad56a0a55adc2df1d3f1e