Check if anyone is using immediate materializer
Summary:
Context:
There are a bunch of optimizations that we want to enable by default in open source, and that would be much easier to do if deferred materializer is the only materializer available. There is also a lot of code we can simplify if deferred materializer is assumed to be the only materializer.

We already deleted virtual buck-out and made deferred materializer the default. This diff checks if anyone is explicity using immediate materializer anywhere in the repo.

Reviewed By: JakobDegen

Differential Revision: D72130733

fbshipit-source-id: 9af46c7e55442c7d1bb7c80e32571ab6e167c4ed