Remove aapt reference from proguard.bzl
Summary:
Aapt is an Android tool and, more importantly, ctx.attrs.ignore_aapt_proguard_config is not a param in non-android build targets. One more Android-specific property which needs to be hoisted.

Rather than imitating the behavior directly I'm following the spirit here and replacing a couple cascading boolean checks with a list of additional proguard configs to tack onto the list created inside the target. This makes it a bit more versatile and less situational.

Reviewed By: IanChilds

Differential Revision: D72085044

fbshipit-source-id: aa8509269baffdea175cf4179730f751cc88c059