Add specific tag for missing targets error
Summary:
`MissingTargetError` is tagged as input which isn't interesting. Add an explicit error tag for it so it can be identified using error tags.

This is especially useful because CI currently uses string matching for compatibility errors, and tags should be used instead as updating error message would also require to update CI's string match

Reviewed By: iguridi

Differential Revision: D72186502

fbshipit-source-id: b135d476a6e668b484a9c2d98d030184ef5a2a01