Add more specific error tag for compatibility errors
Summary:
`TargetIncompatibility` is currently just tagged as `input` which isn't informative. Adding a dedicated tag for it and tag all `CompatibilityErrors` so it is identifiable by error tags.

This is especially useful because CI currently uses string matching for compatibility errors, and tags should be used instead as updating error message would also require to update CI's string match

Reviewed By: christolliday

Differential Revision: D72186503

fbshipit-source-id: 1861dc750ebfb0bf57644c896a5fbd9f08fb9a0c