erlang: share headers with include_only target
Summary:
This modifies how we build application targets so that the regular app target and the `_includes_only` targets share header files.
The header files will be analysed just once - saving on some processing power, but more importantly we won't have to deal with duplicate headers down the line - saving on processing there.

Reviewed By: TD5

Differential Revision: D72183771

fbshipit-source-id: 96c5d4ab9bfa0b510c61315e8d14af6e3db95b44