add test variant to RE action key
Summary:
Adds the optional `variant` to buck2 test passed in from TPX to the beginning of the `action key` suffix. We're trying to run an experiment for Robolectric tests that changes the way tests are ran from all bundled in the same process to all running in separate processes under the same RE action. This results in the two different types of actions having different memory requirements. The `action_key` is meant to help RE decide what resources to allocate to an action and by using the same one here we are not taking advantage of that and causing OOMs.

More context: https://fb.workplace.com/groups/174455863119775/posts/1757984188100260/?comment_id=1763555007543178

Differential Revision: D72194365

fbshipit-source-id: cdcc76781307631727e77647c31044d085845809