Better error for 'run' when target is not in target universe
Summary:
See the test, this error is better.

It'd probably be even better to give the target universe in the error, but it gets moved further up (in the `BuildRequest`) and I didn't think it worth saving just in case we need it for an error message, but open to other opinions. (Also, the error message is reasonably clear without it, IMO).

Reviewed By: iguridi

Differential Revision: D72260769

fbshipit-source-id: 6b893ab37a8c9e7ea242a193d9b2eb2978a28820