Replicate the logic for setDebugAppPackageNames
Summary:
`AdbUtils` is meant to be used with new installer.
In the next diff I have to make it a class - instead of an object and it doesn't worth the effort to wire it up to work with RealAndroidDevice.
All the logics in ReadAndroidDevice is already replicated in AndroidDeviceImpl and it will be removed from our code-base in near future.

Differential Revision: D72261105

fbshipit-source-id: 56ecbf9b89ee13eb078b725b682e627a547fad7b