Delete files in batch
Summary:
Installing exo-package apps, we may need to delete thousands of files (secondary-dexes, native libs) from device.
Currently we are making one `adb` call per each file we want to delete which is not efficiend due to overhead of `adb` call.
In this diff, we write list of files into an arg file and delete files in batch. It is much faster.

Differential Revision: D72261115

fbshipit-source-id: 94fc5b66c26b15cd065b9de63bf47842bc343703