Bulk replace tier0 tags with named tags
Summary:
Since making some IO/Eden error tags non-generic we are now over-aggregating a lot of Tier 0 errors under the same category key.

It's difficult to know which tags overlap with IO/Eden so I made a pass at categorizing them in bulk since we ultimately we should not use ErrorTag::Tier0 at all for normal errors. This only covers about half of the total, will see how many generic errors remain after this.

Used DevMate for part of this but went back and cleaned up most of the names, tried to re-use tags as much as possible to not make tags more specific than necessary.

Reviewed By: Will-MingLun-Li

Differential Revision: D72193984

fbshipit-source-id: ca6322c268b03896b744efcf744ab99bab248515