Don't hand write bxl coerced type determination code
Summary: Instead, just allocate the value and then get its type. This has the disadvantage of being less efficient, but that's not such a big deal because people shouldn't be using this anyway. It has the advantage of being correct, which is a bigger deal. I would write a test demonstrating the incorrectness (some things were reporting dependencies when they should've been labels), but I don't really want to write tests for soft-deprecated functions

Reviewed By: IanChilds

Differential Revision: D72296173

fbshipit-source-id: 1b75ccbf98b945b1f979b481bb73f01274c95a7a