Clear error if min_sdk_version differs between android_instrumentation_apk and the apk-under-test
Summary: The error in the previous diff is really opaque. This `min_sdk_version` difference is the only way that we know it can happen right now, so just provide a decent error for this case (and the more general case that's captured in the previous diff is to prevent future regressions).

Differential Revision: D72076074

fbshipit-source-id: fcad44f655a789783ecb1b2b04a20052a776a1b0