Generate par bootstrap scripts in the preludes
Summary:
For the new live par we still need to generate _bootstrap.sh and the par header file. This diff re-uses the make_par machinery to make this happen.

Note: ideally all of the files would be moved from tools/make_par to buck2/prelude/tools/make_par but there are dependencies on files in tools/make_par.

Reviewed By: christycylee, AishwaryaSivaraman

Differential Revision: D71624608

fbshipit-source-id: d60c35cb43f7c28ec6397e507771248bf50ccf1b