Delete immediate materializer
Summary:
D72130733 adds a soft error to check if immediate materializer is in use and no one is using it.
This diff deletes support for the immediate materializer. This will make it easy for us to clean up code (for example, materializer trait is not necessary if there is only one materializer) and makes it easy for us to enable some settings by default in OSS like `hash_all_commands`.

Reviewed By: JakobDegen

Differential Revision: D72128358

fbshipit-source-id: eabe3787647fe157e8229619bae4261b4859a622