Refactor to use dynamic list of processors
Summary:
To support adding different types of builds, and to avoid code duplication this allows passing a dynamic list of "processors" for handling file entries.

Depending on perf it may be worth refactoring this to just include booleans on the ParInfo and call the processors based on that allowing us to inline the function calls.

Reviewed By: jtbraun

Differential Revision: D72194758

fbshipit-source-id: b0801e2794407abdefa101cf376e3227c6b66a68