xcframeworks | create_modulemap method
Summary:
# Context
Currently, when building an Apple .framework or .xcframework that contains Objective-C, you need to manually create a module.modulemap file and include it.

This is a minor pain in most circumstances, but is a bigger issue when we want to enable the easy production of frameworks for arbitrary targets (e.g. in the '[showcaseable](https://www.internalfb.com/code/fbsource/[e15875b807617ff1f566d108ee6e11568032da4b]/tools/build_defs/apple/showcase/showcaseable.bzl?lines=10-14)' system for IG )

There is already code to generate modulemaps for use in general compilation, this stack extends it to generate working modulemaps for embedding in frameworks.

# This Diff

There is a preprocessor_info_for_modulemap() function that returns preprocessor flags that reference the modulemap, but no explicit artifact for the modulemap.

This diff refactors out a create_modulemap() function that also returns the modulemap artifact for embedding in a framework bundle.

Differential Revision: D72066630

fbshipit-source-id: 370503b5d5b46647c2d4fd0a9d06ca362a9d9ae9