xcframeworks | Implement modulemap auto mode
Summary:
# Context
Currently, when building an Apple .framework or .xcframework that contains Objective-C, you need to manually create a module.modulemap file and include it.

This is a minor pain in most circumstances, but is a bigger issue when we want to enable the easy production of frameworks for arbitrary targets (e.g. in the '[showcaseable](https://www.internalfb.com/code/fbsource/[e15875b807617ff1f566d108ee6e11568032da4b]/tools/build_defs/apple/showcase/showcaseable.bzl?lines=10-14)' system for IG )

There is already code to generate modulemaps for use in general compilation, this stack extends it to generate working modulemaps for embedding in frameworks.

# This Diff

Implements module map generation when the mode is set to "auto" by calling `create_modulemap()`.

We are gathering the tset of `public_framework_headers` for this as this is what we embed in the bundle. There might be further work to make this smarter.

Differential Revision: D72066627

fbshipit-source-id: 1fd081b447ff60e279453273c514d0afcafdd669