Don't set RE use case override on RE client in DICE user data
Summary:
The use case for this RE Client should not be overriden since this is only used in places which do not use the use case from an executor config and should only be used in contexts that use a specific use case.

This client is only used in 3 places:
1. `cas_artifact` action execution
2. Test orchestrator extending digest TTLs
3. Materializing outputs in the installer

Overriding the use case for 1. caused S506044, since changing the use case can cause it to fetch a different artifact in a different bucket which can have a different TTL.

For 2 and 3, the override is not currently useful since the use case override was added to measure resource utilization of eagerbuilds, which should not be running test or install commands.
More generally, because an executor config is not available, 2 and 3 use `buck2.default_remote_execution_use_case` which shouldn't be used if measuring RE resource utilization is important, so it shouldn't be important to override the use case in cases like this.

Reviewed By: IanChilds

Differential Revision: D72427223

fbshipit-source-id: 92441ce660e01bd529b2413b59af0620f3d3249b