Reapply D72432241
Summary:
Reapplying D72432241 which was reverted (D72481517).

Diff was reverted because it broke swift LSP integration.

The LSP requests the `swift-compilation-database` subtarget, and within it, at the end of the json file, are the files to be compiled. This is present from the `argsfile.cmd_form` argument.

In D72432241, we extracted the sources into its own response file (`.swift_files`) and did not put that into `cmd_form`.

Comparison of the `swift-compilation-database` output: https://www.internalfb.com/intern/diffing/?paste_number=1776388850

1. The Swift argsfile should be for the compiler args, not necessarily the whole compile command, so the sources likely never should have been in the `cmd_form` in the first place.
2. This is a bit tricky to fix by adding them to the `cmd_form`, because if we did, then we can't use an argsfile.

To fix this, when forming the compilation database, we can add the swift sources to the output then.

Reviewed By: Wilfred

Differential Revision: D72491919

fbshipit-source-id: 8a8a379f386f96fabe64d27420a35b3b9ddc8b31