Fix RE input when skipping the capturing of incremental outputs
Summary:
RE inputs are non-writeable.

In D72432241, I added a quick workaround to make an input writable so that we could restore the skip incremental outputs option.

In this diff, I fix the logic properly.

The regression happened in D72347143 when we always added the `output_map_artifact` as an input to the command. However, when running with `-skip-incremental-outputs`, we need to write this file in the swift exec wrapper, so we cannot include it in the command line as an input, but rather as an output.

Reviewed By: rmaz

Differential Revision: D72455742

fbshipit-source-id: 976399d92f1f638e4550ee7c33b90bdc4da32234