Allow thin-link and native-link to run in RE
Summary:
We currently force thin-link and native-link to run locally, but they are cheap enough that we can run them over RE. It isn't obvious that this is beneficial though, so gate it behind a config that we can turn on and measure.

My hope is that by allowing these to run on RE, we can get some further build speed wins via caching and reduce memory pressure on the local machine.

Reviewed By: rmaz

Differential Revision: D72357368

fbshipit-source-id: 84736139b69288c20fa81abc270070f99238f986