Associate ManagedRemoteExecutionClient with a single use case
Summary:
Was going to use this to solve the problem in D72427223 before finding a more targeted fix, but it seems worth landing anyway.

This splits ManagedRemoteExecutionClient into UnconfiguredRemoteExecutionClient which needs to have a use case set to be used, and ManagedRemoteExecutionClient which is associated with a single use case and no use case override.

Instead of propagating use cases everywhere and overriding at each call site, override in one place and associate that with the client. This seems preferable and harder to misuse regardless of overrides.

Reviewed By: JakobDegen

Differential Revision: D72424348

fbshipit-source-id: f2abb35eee5801f918e620b388f4b1c7b46fb4ba