Set tasks=false for dep_only_incompatible_version_two soft errors
Summary: `tasks` was set to false to reduce noise because we weren't using the `dep_only_incompatible_version_two` soft error in logview, only in scuba and stderr. Now we want to try using it again given that we are trying another way to increase samples we get in logview, so let's delete `tasks: false`.

Reviewed By: 8Keep

Differential Revision: D72576624

fbshipit-source-id: bc79f532140f9a3b419d3ef458c98c3d794888cb