Prefer remote for cuda preprocessing actions
Summary:
The preprocessor sometimes hardcodes the cwd in the .ii output files like this:
```
# 0 "fbcode/caffe2/aten/src/ATen/native/cuda/ActivationLeakyReluKernel.cu"
# 1 "/data/users/zhuhan/fbsource//"
# 0 "<built-in>"
# 0 "<command-line>"
# 1 "fbcode/third-party-buck/platform010/build/cuda/12.4/include/cuda_runtime.h" 1 3 4
# 61 "fbcode/third-party-buck/platform010/build/cuda/12.4/include/cuda_runtime.h" 3 4
#pragma GCC diagnostic push
...
```
This is another factor that is different between local and remote builds, and that hurts cache usage. I'm not exactly sure how this path is used. But editing this directly is too hacky. Let's just add `prefer_remote` to the preprocessing actions to disable concurrent execution. The remote action should succeed and in the rare event that it doesn't, the action will fall back to local execution.

Reviewed By: wenxin0319

Differential Revision: D72489323

fbshipit-source-id: f735727a0471b71bd335480b30ed9335cec70720