Move streaming logic out of BuckSubcommand
Summary:
This stack unifies logging code between streaming and non-streaming commands. This fixes some discrepancies in logging (command report) and enables later cleanup to (almost) always log error messages when commands fail.

As it stands `BuckSubcommand` is useless since every `BuckSubcommand` is also a `StreamingCommand`, this makes it into a common interface for all commands.

Currently subscribers which handle logging are created and finalized in `BuckSubcommand` for streaming commands, or in `ClientCommandContext::instant_command..` methods for non-streaming commands.

This and the next diff make it so that `ClientCommandContext::exec/exec_async` are responsible for creating and finalizing subscribers, and `BuckSubcommand::exec_impl` handles command specific logic.

At the end of this stack, non-BuckSubcommand command types where `exec` is called directly (not via `ctx.exec`) should only be used for commands that either do no logging or delegate to a more specific `BuckSubcommand`.

Reviewed By: JakobDegen

Differential Revision: D69029676

fbshipit-source-id: 8b53385a4301b77cc0d77c2e94c1af3fccd90f33