Do not use runtime manifest for native python
Summary:
We previously used the runtime-manifest to tell make_par to set the binary entrypoint to the native-python binary. With this change we now pass the binary name directly and no longer generate a runtime manifest unless we are actually bundling the runtime.

Note: This changes the manifest to no longer include the "bundled-runtime" attribute for cpython + native python.

Reviewed By: zsol

Differential Revision: D72480889

fbshipit-source-id: fa10466756a26f00ab8f45475d8b23c40ad360a3