Add parser support genrule()
Summary:
Currently we only support genrule() as part of native rules. This adds it as a default_global rule as well.

User report: https://fb.workplace.com/groups/autodeps.users/permalink/2942180399283773/

For some reason, the native.read_config function is erroring out in Autodeps (not sure why, it should be supported). It's not really necessary and beyond the scope here, so just removing it from the file in question to unblock.

Reviewed By: IanChilds

Differential Revision: D72593265

fbshipit-source-id: 59c759987795bf78205a5fed780be36d319c4c69