Choose proguard version by toolchain
Summary:
In D72026788 I removed the proguard jar from the android toolchain since it could be fetched from the java toolchain instead, being more generic. But it turns out the Android version of proguard is too old to compile against Java 17 targets, so I have to choose the proguard version based on the toolchain.

This diff returns the proguard_jar field to the android_toolchain, hoists the property out of proguard.bzl and allows the binary rule to fill in which proguard version to run on the input libraries.

In this way we can continue to run Android's proguard 4.7 on Android binaries, while using the newer Proguard 7.2.2 for Java binaries.

Reviewed By: IanChilds

Differential Revision: D72591670

fbshipit-source-id: dfb576bcc3ecd9f26181b5b250e91481dfe150b4