Use BuckSubcommand for clean
Summary:
This is a bit complicated since `clean` is a client only command and `clean-stale` is a streaming command. I was going to change the API to `buck2 clean-stale` to make it a separate command and simplify this but a few people are using `clean --stale` via the CLI so leaving it for now.

Note that switching this causes `buck2 clean` to write a command report so it will leave behind a log dir with a single entry. Updated tests to expect this.

Reviewed By: JakobDegen

Differential Revision: D72545588

fbshipit-source-id: a454e82843a36286db6ea599b69d469fa276ea2d