Emit incremental metadata from rule layer
Summary:
### Context
We are exploring the integration of the Kotlin Incremental Compiler into Buck for the incremental compilation of Kotlin source files. The goal is to improve build speed by only recompiling modified source files within a target, rather than rebuilding the entire target as is currently done.

### This stack
This stack aims to implement a mechanism that forces KIC to run non-incrementally in specific scenarios, such as when breaking changes are made to certain metadata files.

### This diff
**Implementing Incremental Metadata Generation**
- Creates incremental_metadata.proto.json file
- Passes the generated file as input to the Java layer

----
**Internal**
[Doc](https://docs.google.com/document/d/1-zhGPkmt3I2os-kcIaFk9SvT6odt2AJWswSY-J_6AWs/edit?usp=sharing)

Reviewed By: navidqar

Differential Revision: D72721609

fbshipit-source-id: cc66e509242dc9a79858b5162100507581884a81