Decompose packaging_jar_args to package map to prep for proguard support
Summary:
To prep for D72651259, where we will introduce proguard, I have split out the refactor to java_binary to make room for the proguarded jars path.

We will need a mapping of original jars to their labels for proguarded incremental binaries (so we can look them up by their proguarded jars), so I've extracted that logic here. Check the next diff for the final implementation with proguard.

Reviewed By: IanChilds

Differential Revision: D72651259

fbshipit-source-id: 46413968719913ef1dcc104ed52b086b6b01b20c