Add test for hg revision in invocation record
Summary: Add a test for checking hg commands are called properly and reported to invocation record. Manually enforced the test to be 5+ seconds long to ensure the hg commands finish, which should be enough in most cases. This is non-deterministic and could be an issue though if there are a lot of local changes and `hg status` takes way longer, which will result in failure. I'm fine with not including this if anything thinks that's an issue

Reviewed By: iguridi

Differential Revision: D72621743

fbshipit-source-id: 3ebd5eb3bdc19cf4ee4e698621d70f4d63749fc4