Add platform's scripting language to OsLookup provider
Summary:
This is one of the essential characteristics of the exec OS that code in the prelude needs to know about, in order to do things like `-Clinker` flags with a script that contains linker arguments.

Used in D72804575 for cleaning up the canonical invocation of `cmd_script`.

```lang=diff
 cmd_script(
     ...
-    os = ScriptOs("windows" if ctx.attrs._exec_os_type[OsLookup].platform == "windows" else "unix"),
+    language = ctx.attrs._exec_os_type[OsLookup].script,
 )
```

Reviewed By: capickett

Differential Revision: D72804576

fbshipit-source-id: 707d11eb01eb514179eaee024d292645232ffa9d