Change base dep label comparison back to raw_target() for incremental jar base dep comparison
Summary:
Swapping back to the comparison behavior we had before D72651259. Tests were passing there as well, so this is probably functionally equivalent, but better safe than sorry. Using label's raw_target instead of the whole label for the base dep comparison.

The left hand side stays the same because we already extract the raw_target from the label when we define jars_to_owners, which is what packaging_dep_infos is.

Reviewed By: IanChilds

Differential Revision: D72797170

fbshipit-source-id: a3a019b791d00b164cdaced34df2ad5bc027c22f