Replace RebuildReason sealed interface by enum
Summary:
### Context
We are exploring the integration of the Kotlin Incremental Compiler into Buck for the incremental compilation of Kotlin source files. The goal is to improve build speed by only recompiling modified source files within a target, rather than rebuilding the entire target as is currently done.

### This stack
This stack aims to implement a mechanism that forces KIC to run non-incrementally in specific scenarios, such as when breaking changes are made to certain metadata files.

### This diff
I replaced RebuildReason sealed interface by enum since I think added complexity is not worth a precise log message.
----
**Internal**
[Doc](https://docs.google.com/document/d/1-zhGPkmt3I2os-kcIaFk9SvT6odt2AJWswSY-J_6AWs/edit?usp=sharing)

Differential Revision: D72787107

fbshipit-source-id: 760b73fbaa8a17261900095ad6109b0bf9d0f5ec