Disable source changes detection
Summary:
## Context
We are exploring the integration of the Kotlin Incremental Compiler into Buck for the incremental compilation of Kotlin source files. The goal is to improve build speed by only recompiling modified source files within a target, rather than rebuilding the entire target as is currently done.

Currently, we allow the compiler to calculate changed files within a target, even though this information is available from Buck as an action input. We can reuse the information from Buck and skip one compiler phase entirely.

### This diff
Disables source changes detection by Buck, as it was accidentally pushed with D72046347 and has not been thoroughly tested in all scenarios yet.

Reviewed By: navidqar

Differential Revision: D72965585

fbshipit-source-id: 876628b4ca9c560ed7427b3e92137b1157d26f37