disable test_version_control_collector for mac
Summary:
The test's never worked on Mac since it was committed in D72621743 https://www.internalfb.com/intern/test/562950162346504.
The reason for this that we use `symlink` redirection for test on Mac.
So `hg root` invocation happens not inside repository:
```
abort: '/Users/yurysamkevich/.scratch/UsersZyurysamkevichZ.scratchZfd038bb56980958fZtmpr7q22aopZproject/edenfsZredirectionsZbuck-out/v2' is not inside a repository, but this command requires a repository!
```

Reviewed By: Will-MingLun-Li

Differential Revision: D72971788

fbshipit-source-id: 3f77be4e6e388eb859fe6eccb04cd0c13ca54b67