Specify `-fmodule-file-home-is-cwd` to the module builds.
Summary:
This diff specifies `-fmodule-file-home-is-cwd` flag which allows to produce relocatable PCMs without the internal patch from T32246672.

In the current state and our use of Clang, this flag ends up just being ignored. However, in the rollout of Clang with the internal patch reverted and [llvm#135147](https://github.com/llvm/llvm-project/pull/135147) backported, it will end up having the same effect of storing relative paths in PCMs as before.

Reviewed By: dmpolukhin

Differential Revision: D72999617

fbshipit-source-id: 222857daebed514951c72882e46b0c6478938544