Fix deprecated_config to only read from corresponding cell's buckconfig
Summary: `deprecated_config` is used to fire soft errors on `read_config`s of specific buckconfig keys. Currently, it is specified under the `deprecated_config` buckconfig key in a per-cell basis but falls back to the root `deprecated_config` section if one isn't specified in cell. However, this means that it's not possible to instrument just the root cell. This diff removes the fallback to the root cell behavior.

Reviewed By: mzlee

Differential Revision: D73057614

fbshipit-source-id: 4525d1c7d4d2ea69e45ae098615bbd7f48fe3795