Avoid unnesesary configuration transitions
Summary:
We introduced a configuration transition for forcing windows builds to use mingw toolchain.
Accidentally this transition also has been applied to Linux builds causing dependencies of the same C++ binary to be linked twice.
The issue was that `prelude//os/constraints:os` was actually undefined for Meta-internal target platforms.
The `ovr_config` and `prelude` are both aliased as `config` so now transition works for both builds OSS and internal

Reviewed By: echistyakov

Differential Revision: D72975131

fbshipit-source-id: 20bd828977e6a2570a0bd61bf7c255912352ad0b