expose configured target graph sketch to build report
Summary:
Like it says in the title. The end goal is to expose this to Citadel so that it
may report

This is spiritually Clucky
(https://fb.workplace.com/groups/1016632666864038/permalink/1041895137671124/),
but done in a way that can be hooked to Citadel a little more easily and reuses
traversals that we already do.

Reviewed By: scottcao

Differential Revision: D73034142

fbshipit-source-id: bba346990387ce7e50d327e7d62a88534f37c333