Use installer timeout for file_ready requests
Summary:
We've had user reports of the installer being stuck indefinitely while sending files which I suspect are either due to installing over the network with a poor connection, or perhaps because it's blocked on the user taking some action on the device.

Adding a timeout to all requests made from the installer over GRPC, although send_file is the one most likely to be a problem since other requests don't currently talk to a device.

Setting the timeout for send requests to 5 minutes for now since there are a significant number of long duration installs, although it would be a good idea to reduce this.

Reviewed By: JakobDegen

Differential Revision: D72996131

fbshipit-source-id: b0ef95e69c8103541224fa4b15392fded6c48d6e