remove config read
Summary:
Six months after the fact we got rid of the todo, pretty good time if im hoenst.

This adds a bunch of stuff to the transition that basically implements 'check if the user requested any sanitizers, if so don't let opt-by-default apply'
In the future we can probably do something where we only don't override the sanitizers if the user has specified it but we're too close to GA rollout to change the behaviour right now.

The transition implementation has (AFAICT) the exact same behaviour as the previous impl w/ the build mode name checks. Verified by running the tests on new code.

Reviewed By: zsol, scottcao

Differential Revision: D72971232

fbshipit-source-id: 9f8e4aa70bfcdfd444338971c14a4644da203b99