add the prelude logic
Summary:
This will inject a specific list test command (when the configs are present) for generating the list test comand for android robolectric tests.

The next diff will fully test this e2e.

Notable configs:
`-c testinfra.build_codesense`

If true we will build the codesense binary (for development)
if not true we will use the dotslash version

 ---
`-c testinfra.codesense_static_listing`
If true we will inject the above RunInfo and use it in the robolectric.bzl
If not true we will not have this injected and the existing path will be used

## Rollout plan
For phase #1 we will create shadow runs for the robolectric continuous runs and inject the config `-c testinfra.codesense_static_listing` for those shadow runs. We can then compare between both runs and ensure the same tests were discovered.

After we have parity and all of the targets that won't support this (e.g., Parametrized tests)  we can move to a label based opt-out.

Reviewed By: jselbo

Differential Revision: D73158309

fbshipit-source-id: 179d6a8866c05123a29be0aaf32804afcb305a08