command alias: Start a `tests/prelude`
Summary:
The way we should be writing prelude tests is by defining a bunch of targets that are "tested" by building them. They're easier to maintain, cheaper, more td correct, and faster to run compared to the existing e2e test crap.

We've been talking about this for too long, I think it's time we start. This diff sets up a `buck2/tests/prelude` with a pretty straightforward citadel config and moves the `command_alias` tests in there.

There's lots left undone in this diff, but honestly even this simple thing is enough for many use cases. Also, I don't think it's a viable strategy to say that we're never going to do this until someone has the time to make it 100% perfect. So I'm starting, and we can iterate

Reviewed By: IanChilds

Differential Revision: D73173418

fbshipit-source-id: ca3d6e700e99ae859eb1bc5ef39217fa77827ddd