Swap redirect_stdout.py arguments to support nontrivial RunInfo
Summary:
Similar issue as {D73323527}. The `llvm_lines_tool` field of the RustToolchainInfo is in general a `RunInfo`, which does not necessarily get rendered to a single argument.

https://www.internalfb.com/code/fbsource/[156ee3790cc93cc9ea63164bc0e62b6f369df817]/fbcode/buck2/prelude/rust/rust_toolchain.bzl?lines=131

Redirect_stdout.py, which is used by the `…[profile][llvm_lines]` subtargets, was expecting argument 1 to be the `llvm_lines_tool` and argument 2 to be an output file, followed by additional flags. However, argument 2 might be part of a multi-word `llvm_lines_tool`.

Swapping argument 1 and 2 is an easy fix because `llvm_lines.as_output()` is definitely going to be a single argument.

Reviewed By: JakobDegen

Differential Revision: D73324976

fbshipit-source-id: b88a816c39077d65a86f32abe28635dc05ac7d68