command alias: Remove `sed` step on Linux
Summary:
The previous diff explained what this was doing

In this diff, replace it with a different technique - instead of doing the replacements with `sed` at build time, do it in bash at runtime.

Other than being easier to implement and more correct, the primary reason that this is better is because doing `sed` at build time requires us to implement this twice depending on the exec platform, while at build time we can afford to do it only once for the target platform. As mentioned, in practice that wasn't implemented on Windows, and so this fixes a bug in that case, though unfortunately that bug is hidden behind [this other one](https://fburl.com/code/nqw0nlx5).

There's another reason this is an improvement: With the sed-at-build-time technique, we ran a high risk that command_alias outputted inconsistent scripts depending on its exec platform (something that actions generally shouldn't do). By not doing anything on the exec platform, we know that's not the case

cc ndmitchell krallin : There's a comment here saying you discussed this, was this option considered and rejected for any reason I'm not thinking of?

Reviewed By: dtolnay

Differential Revision: D73292917

fbshipit-source-id: 619cf8b4b6e1652de0b770805ac4848781b70b74