Allow parsing fully relative imports based on buckconfig
Summary:
We decided to use `fbsource/.buckconfig` to set the allowlist for directories allowed to use relative paths.

Here, I'm passing the buckconfig object down to `InterpreterForCell` and using it to get `allow_relative_paths` depending on the root buckconfig and the `import` path.

I'm not sure at all that this way of doing things is correct, so I appreciate any comments with better ideas.

Reviewed By: JakobDegen

Differential Revision: D72739365

fbshipit-source-id: 5dc7db79c10e12e3d02b8189d2360befe1314fca