Always disable coverage for test_main_gen
Summary:
`test_main_gen` is an on-the-fly generated test runner.
It's invisible to the user and is only needed to run unittests.
https://www.internalfb.com/code/fbsource/[85cbc18e3a76]/fbcode/buck2/prelude/go_bootstrap/tools/go/testmaingen.go?lines=415

There is absolutely no need to have coverage enabled for this file - it will only slow down test execution (even though negligibly).

However, this change unblocks an important downstream fix - for packages with name overrides.

Reviewed By: podtserkovskiy

Differential Revision: D73385887

fbshipit-source-id: 81108d447a6d25850923d292df5f4c9332fd33b9