Write client error to invocation record from ExitResult
Summary:
Client side errors that occur either too early or too late are not logged to the invocation record because they happen outside of a few stream processing methods in the client.

We can now log these from `ClientCommandContext::exec_async` which is much closer to the top (or bottom?) of the stack for command execution on the client. Only a small amount of code runs after this to finalize logging which should not be able to cause commands to fail.

Since only a single client error can be propagated via the exit result, this also replaces the list of client_errors in the InvocationRecorder with a single optional client_error.
Later diffs in this stack should help to validate that we aren't losing information by reporting a single error here. This diff was initially at the end of this stack, but it turned out to be much simpler to fix this part first and clean it up after.

Most of the rest of this stack is refactoring and clean up to fix some other discrepancies in error logging to make this less fragile.

2 other small fixes required by this change:
- We shouldn't report an extra error from `handle_daemon_connection_failure` since this is now captured by the exit result and would otherwise be a duplicate error.
- `buck2 debug crash` should always return a non-successful exit result. If a command is successful we shouldn't be logging any errors, this only worked before because `handle_daemon_connection_failure` errors were propagated out of band.

Reviewed By: Will-MingLun-Li

Differential Revision: D73326401

fbshipit-source-id: 3626240c54bb4a378df9e8e5ba109fd24385dff2