Return a single error from daemon on server failure
Summary:
This command error is only used to report a generic error in the daemon which is never a list of errors and shouldn't be either.

We should only produce lists of errors from commands doing parallelized work (analysis, actions etc.) that is relevant to the user. Those could be generalized but are currently returned from per-command `response` values.

Reviewed By: Will-MingLun-Li

Differential Revision: D73263990

fbshipit-source-id: 8c900d4450b2cf6b85360f6150117d744bba98a9