Set all invocation record errors from exit result
Summary:
Ensures the errors in the command report and the invocation record are always the same, with the one exception being the ClientGrpc errors which are augmented in the invocation record (see FIXME).

This means if we hit errors on the client side we no longer report action errors to the invocation record, only the client error, which isn't ideal but should be very rare and should be fixed by not throwing errors at that stage and not overriding the exit result, instead of having special handling in the invocation recorder.

Reviewed By: Will-MingLun-Li

Differential Revision: D73335129

fbshipit-source-id: ee87dd8147c49746f2a89535ed41fa3d7cfb01c7