Finalize test command errors in daemon
Summary:
The purpose of this is the change to `handle_command_result` in the invocation record, so that we aren't modifying command errors only for the invocation record and we always have the same errors in the command report and the invocation record, which is used in the next diff.

Also in general we should have as little logic in command specific client side code as possible.

Reviewed By: Will-MingLun-Li

Differential Revision: D73335127

fbshipit-source-id: ac5cd90b2844e5f2556d1bd9f5f5fd24ec4ebd83