Write errors that haven't been emitted to console from handle_exit_result
Summary:
At this point there are 3 subscriber methods that trigger a console to 'finalize': `handle_command_result`, `handle_error` and `handle_exit_result`.

There are also 3 places where we write command failing errors to the console:
- superconsole (`StatefulSuperConsoleImpl::handle_command_result`)
- errorconsole (`ErrorConsole::handle_command_result`)
- report method before client terminates (`ExitResultVariant::report`).

Currently daemon infra errors are written to the superconsole or errorconsole depending on console configuration and 'special' client side errors are written by `report`.

The next diff causes daemon infra errors and client side errors to be handled the same. Without this change, both would be formatted using `report`. With this change, both are formatted using the superconsole/errorconsole and only fallback to use `report` if that fails.

The added logic in this diff is copied from other places, but is removed in later diffs in this stack.
- The mechanism in `handle_console_write` is essentially taken from `events_ctx.handle_error_owned`
- The added logic in superconsole/errorconsole is from `handle_command_result` and `handle_error`

It would be good to clean up the console formatting code more (can we just use `report` and get rid of the errorconsole and FinalizedSuperConsole?), but this mechanism for marking within the exit result as emitted would still be needed and it wouldn't hugely simplify this code.

Reviewed By: Will-MingLun-Li

Differential Revision: D73326400

fbshipit-source-id: 8aa8569e4e2d9abc68afea33bca2e1a2684073f5