Separate emitted errors in the exit result
Summary:
Clarify which errors in ExitResult are emitted command result errors, and make it more clear that the error in the ExitResultVariant hasn't been emitted, and combine the two when creating the command report.

This enables refactoring in later diffs to simplify emitting to the console and to re-use these errors for both the command report and the invocation record.

Reviewed By: Will-MingLun-Li

Differential Revision: D73268313

fbshipit-source-id: ef275b71b22d96d99eaaf64adf7429e69508dcaf