Record build error count for test commands in daemon
Summary:
This and the next diff make the errors generated by test commands match the same structure as every other command, so that we don't have different errors in the command report and the invocation record.

Currently we add additional errors to the invocation record that are not build errors. If we change the errors sent in the `TestResponse` we need to send the build error count separately to get the correct count of build errors on the client.

Reviewed By: Will-MingLun-Li

Differential Revision: D73335128

fbshipit-source-id: c83d2369b1d5185c924231ff0a5a130d013bcdfa