Set exit code based on error category in ExitResult::err
Summary: If a command fails with ExitResult::err the error category should affect the exit code. Currently this is used rarely and mostly for infra errors so it's not clear how many incorrect exit codes this will fix, but this is needed for the next diff which uses ExitResult::err for more common errors.

Reviewed By: Will-MingLun-Li

Differential Revision: D73452566

fbshipit-source-id: 0d8a12ba287a7917148115986a10bd7e84b13a2a