Add DetailedAggregatedMetrics instant event
Summary:
This event will hold data that we aggregate over the full target and
action graphs for each top-level target (and for all of them together).

For the top-level targets, we will compute both full sums of metrics and
amortized versions of those metrics where each metric is split across all the
top-level targets that have the associated node in their graph.

Reviewed By: JakobDegen

Differential Revision: D73392171

fbshipit-source-id: 2908cc9cd0cb1e05fb4438ad3393b432f156c39f