Support buildscript_run with non-default name for manifest dir target
Summary:
For example:

```lang=starlark
buildscript_run(
    name = "serde-1.0.219-build-script-run",
    package_name = "serde",
    buildscript_rule = ":serde-1.0.219-build-script-build",
    features = ["default", "std"],
    manifest_dir = ":serde-643d78919f35d883.git[serde]",
    version = "1.0.219",
)
```

This makes it possible to run using a subdirectory of a `git_fetch` repo as the working directory for a build script. The git repo part of PR D73389843 only worked for crates that were located at the root of their git repo, which is not always the case.

Reviewed By: diliop

Differential Revision: D73601141

fbshipit-source-id: c0a75474e7e6e70afe5fda9ddfc2c835ffa650e3