Remove filegroup construction from cargo.rust_binary
Summary:
D73389843 changed `cargo.rust_binary` to recognize the case when the Rust binary was really a build script (target name ending in `-build-script-build`) and used some of the attributes of `cargo.rust_binary` to produce a directory artifact to be used as the current directory by the corresponding build script execution target (name ending in `-build-script-run`).

D73601139 improved the git dependency case to use the correct subdirectory of the git repo as the current directory for build script execution, instead of the repo root, to align with Cargo's behavior. It directly informs the build script execution target what directory to use, using a sub-target of the `git_fetch` target.

```lang=starlark
buildscript_run(
    ...
    manifest_dir = ":serde-643d78919f35d883.git[serde]",
)
```

and D73625775 improved the local path dependency case to directly inform the build script execution target what local directory to use also.

```lang=starlark
buildscript_run(
    ...
    local_manifest_dir = "sources/with_buildscript",
)
```

In both cases, we now no longer rely on `cargo.rust_binary` to reverse engineer the filegroup that represents the crate's manifest directory, so this diff takes out the logic from D73389843.

Reviewed By: JakobDegen

Differential Revision: D73635117

fbshipit-source-id: 46002a3895a9d15ad30f742769c700bb61294c56