Leverage Buck Incremental Action Input
Summary:
### Context
We are exploring the integration of the Kotlin Incremental Compiler into Buck for the incremental compilation of Kotlin source files. The goal is to improve build speed by only recompiling modified source files within a target, rather than rebuilding the entire target as is currently done.

Currently, we allow the compiler to calculate changed files within a target, even though this information is available from Buck as an action input. We can reuse the information from Buck and skip one compiler phase entirely.

### This diff
Allows for reusing information from Buck.

Differential Revision: D73599076

fbshipit-source-id: eacf478ddf2b107bafa87d72eb823bb04ef08401