rust-project: Explicitly log progress
Summary: Currently if rust-project emits a malformed `DiscoverProjectData`, rust-analyzer will silently ignore it. I want to change this in https://github.com/rust-lang/rust-analyzer/pull/19684, but we need to first ensure that rust-project only emits these events and not other logs. Currently, all tracing events are emitted by the `ProgressLayer`. This diff makes it so only events with an explicit `kind` are emitted.

Reviewed By: davidbarsky, dtolnay

Differential Revision: D73638698

fbshipit-source-id: 11cc8d95f250ab8fa3d315a2ac69ad01e2f7372d