use unique inodes for swift and swiftmodule compile action
Summary:
We require unique inodes to compile actions to ensure deterministic output. Clang uses an inode based file cache, so de-duplicating identical headers to the same inode can cause header resolution to flap depending on import order. This can cause build failure when headers are not the expected size for precompiled clang modules.

This was previously enabled only for PCM compilation, but is also required for Swift actions that produce headers.

Reviewed By: chatura-atapattu

Differential Revision: D73668727

fbshipit-source-id: 9234058c0cb09505997242c4634dd6efca1405ed