Write invocation outcome to invocation record
Summary:
Currently `outcome` in buck2_builds is calculated in ingress, using this sequence of checks:
- If `command_end.is_success` (or `instant_command_is_success`) is present, command is marked as `Success` or `Failed`
- If the equivalent of `crashed` in this diff is true, command is marked as `Crashed`
- If `!has_end_of_stream`, command is marked as `Cancelled`
- else command is `Unknown`

Basically, ingress expects to receive `command_end` and if it doesn't have it, tries to guess the reason the command failed from other events.

The main motivation for this change is to reduce instances of `outcome=Unknown`, this should make it so that commands that fail on the client side but aren't detected as crashes are no longer `Unknown`. This gives `Unknown` a more specific meaning which is any case where errors do not match the exit code, which should be possible to eliminate.

This should also be a better way of detecting cancellation which should work for non-streaming commands, and allows cleaning up `command_end.is_success` and `instant_command_is_success`.

Reviewed By: Will-MingLun-Li

Differential Revision: D73556164

fbshipit-source-id: 9ccccac50f7aaff266ca87e00466d01625c8f5fc