Force metal/metalib to only try to execute on macOS
Summary:
Mark the genrules that use `xcrun` to execute the `metal` compiler and the
`metallib` tool as only compatible with macOS. This allows them to be shipped
to macOS RE workers even if the build is started from Linux.

This should be replaced by the work in D71829183 when ready.

Needed to remove the tag `uses_xcrun` because it marks the actions as local
only, which makes it impossible to send them from a Linux machine to a macOS RE
worker.

Reviewed By: IanChilds

Differential Revision: D73812180

fbshipit-source-id: 990dd23b26ccca8741264bce9eec4b5626c2af40