erlang: use write_json(with_inputs = True)
Summary:
This uses a feature of buck2 to output a json file as an object that automatically tracks dependencies, so we don't need to manually set them as `hidden` on `cmd_args`.

This significantly cuts down on memory use.
On a full run, the retained memory is lower by 780MB (or 12.5%) and total allocations are lower by 1.9GB (or 13%).
Measurements done using `buck2 profile analysis --mode=heap-summary-allocated -o heap-summary waserver//...`

Reviewed By: TheGeorge

Differential Revision: D73922128

fbshipit-source-id: a940d3f6908f8fb5af66d2c091360ec6604d61c2