module: Make functions and methods work with generics
Summary:
Like it says in the title, at this point just thread it through

Unfortunately, this whole thing with generics is not actually going to work out. I wanted to use this for something like an alternative to the `internal_provider` macro (though not on providers, but enums instead) - currently, we use a macro to generate a `#[starlark_module]`, but I instead wanted to write something like

```
impl<P: InternalProvider> StarlarkValue<'v> for P
```

and then generate impls of `InternalProvider`.

The problem is that even with the things so far done, it turns out that getting the lifetimes right across the type and trait definitions, `StarlarkValue` impl, and `MethodsBuilder`, is really, really hard. I got to the point where I was pretty close, but am now calling it because:

 1. The amount of type system knowledge needed to be able to maintain the result is extreme
 2. Some of the patterns I was using were getting pretty close to known unsoundnesses in the language, specifically #84366

This is not without hope though - actually doing the work of getting rid of the `Value`/`FrozenValue` impls would make this a lot more feasible because we wouldn't have the `V` flying around too, saving a bunch of HRTBs and such things

I'm submitting the stack anyway since, on balance, I think it's probably better to have than not

Reviewed By: Nero5023

Differential Revision: D73726001

fbshipit-source-id: 014678dbdfc18a2ef5404cd7bc5548521a2a104a