Don't set should_restart on restarted builds
Summary:
This is used to filter failed commands that should have been restarted, so that we only count the result of the subsequent restart command, however sometimes `should_restart` is true for the restart command, which won't restart again.

If `restarted_trace_id` is set we won't restart so `should_restart` should not be true.

This can be worked around in scuba by checking both columns but it shouldn't be necessary to do that.

Reviewed By: iguridi

Differential Revision: D74026546

fbshipit-source-id: 4cf8a04810a2bdd9d99faef4eec787dc2e832fb2