Add preemptible to invocation record
Summary:
This can be determined by checking argv, but it's error prone to do it that way, plus I want to filter by this in logview where we shouldn't store argv for every sample.

This will always be UNSPECIFIED for non-streaming commands, not clear if it can be set and used by some non-streaming commands that connect to the daemon but would require a bunch of refactoring to fix and doesn't seem critical for those commands.

Reviewed By: iguridi

Differential Revision: D74026545

fbshipit-source-id: ff4bd4ab41d264a5f3b85ae8a655d899ad0a2c36