Fix command report writing with broken buckconfig
Summary:
The command report test in the buck wrapper that tests that command report logging from the wrapper which happens to use a minimal repo with an empty buckconfig.
A recent health check change broke this because the client to exit early if `daemon_startup_config()` returns an error. Changing this to just disable health check if the config isn't valid.

The fact that subscriber initialization returns a `Result` that fails the command immediately is still a footgun, planning to clean that up in a later diff.

Reviewed By: iguridi

Differential Revision: D74089586

fbshipit-source-id: cc3dd93ca9b4f675003a22f6e2f079ecd15a8528