Implement computation of detailed aggregated metrics.
Summary:
This ties together the pieces put in place for computing and reporting detailed aggregated metrics.

We track analysis events so that we can traverse the action graph (across all
of tsets, actions, dynamic outputs) and track execution events so that we can
compute metrics across the whole graph.

To compute metrics, we traverse the action graphs associated with each
top-level target to get a mapping of `actionkey->[topleveltarget]`. This allows
to aggregate per-action metrics to those top-level targets and to compute an
"amortized" form of that same aggregation (where the cost is divided equally
across all the topleveltargets that require it).

We do similar for some per-analysis metrics (by traversing the configured graph).

Reviewed By: JakobDegen

Differential Revision: D72591446

fbshipit-source-id: ae09936e753b199661a86c57190202e7d0274b33