Cache exec_is_windows and path_sep in CompileContext
Summary:
The definition of these is simple enough:

https://www.internalfb.com/code/fbsource/[bf99b8f7bef01675f767752e8c852a7f849de96a]/fbcode/buck2/prelude/rust/build.bzl?lines=998-999

but they are used pervasively in build.bzl, and it's annoying to either repeat the definitions (6 times for `exec_is_windows`, 4 times for `path_sep`) or pass them around everywhere they are needed (like for `process_env`).

There is already a context object that we pass around to most of the functions.

Reviewed By: JakobDegen

Differential Revision: D74132457

fbshipit-source-id: 703ff0070f540594ba2fd3d72b7a0fefe333821c