Improve remap-path-prefix for Rust targets near repo root
Summary:
`ctx.label.path` (StarlarkCellPath) has the inconvenient behavior that when interpolated into a cmd_args, if the label's package has fewer than two components, it gets an extra `./` prepended. This affected all Rust compiler diagnostics in Rust targets in the project root and inside of top-level directories.

Maybe this can be fixed in Buck but for now work around it in the Rust rules.

Context:

> D46842167 did this intentionally
> I'm not sure if I think that was right. I think it's probably better to try and do this just for arg0 in the remote executor like has been done in D46440477 for local executions? Not sure

Reviewed By: JakobDegen

Differential Revision: D74132455

fbshipit-source-id: 098f8233275b5ea1f78d07943918fe9879717d3c