Add CommandEnd.build_result
Summary:
I want to get rid of CommandEnd.is_success, however there is a possibly valid use case for it when calculating `min_builds_since_rebase`, which is to report if a target finished building, independently from if a command finished successfully, although it's not totally correct at the moment.

Since `is_success` is used when calculating `min_builds_since_rebase` it would be a minor regression if we used the exit result instead of `is_success` since fewer incremental builds would be detected. We should still consider a build incremental if the preceding command built successfully but only failed after building.

The command this may be most relevant to is install which doesn't actually work correctly but it's not trivial to implement so I left a TODO. I fixed the implementation for `test` commands only because it's easy.

The result of this is that `CommandEnd.is_success` can be removed in the next diff, while making it clear why we might not want to use a generic success value in this case.

Reviewed By: Will-MingLun-Li

Differential Revision: D73794593

fbshipit-source-id: 36f46e95336d0b528a94f10b4ccf138f904d6ff9