Fix broken editable python wheel
Summary: editable wheel is broken due to P1803526873. Looks like certain module dirs do not exist. The root cause is that when creating missing __init__.py, we are not creating it under the right parent path (that should start with buck-out/...). It uses the package name to create, which is incorrect. This diff saves the package dirpath along with the names, which should fix the issue.

Reviewed By: andrewjcg

Differential Revision: D74186377

fbshipit-source-id: 6b7211884918ba002327fb7cb80a960d1d1045fe