Don't return error from get_console_with_root
Summary:
If initializing the superconsole fails, we should fall back to the simple console and continue instead of failing the command immediately and breaking other logging.

Note this doesn't pass the health check receiver to the simpleconsole if super console initialization fails since we can't just clone the receiver, it's not trivial to fix but doesn't seem critical.

Reviewed By: iguridi

Differential Revision: D74108643

fbshipit-source-id: 304a44f775c37e2585af8845ea6536d0860d75ea