Don't return an error during subscriber initialization if InvocationPaths are not available
Summary:
If initializing InvocationPaths fails, it is likely the build will fail, but we should not exit immediately since we should still try to write an invocation record.
If this happens we should just disable the subscribers that have a hard requirement on this, which are the event log and re log subscribers.

Reviewed By: iguridi

Differential Revision: D74108642

fbshipit-source-id: 1f2d59fc768e7a5b54d9fa126f02d331c1b51ee5