Don't return error from subscriber init methods that can't fail
Summary:
The rest of this stack causes subscriber initialization to not return `Result` so that we don't exit early and break all logging if there is an error, instead we should only disable the subscriber/subscriber features that failed to initialize.

As a first pass, this diff makes code that already doesn't throw any errors not return `Result`.

Reviewed By: iguridi

Differential Revision: D74108641

fbshipit-source-id: 780a75a0e01e0565c70d007c800dc9556b6119c3