Do not include dependent artifacts with extracted manifest
Summary: When we build inplace pars we need to make sure the manifest does not pull in any dependent artifacts or the action won't run until those artifacts are built and materialized. This allows bytecode compilation and inplace par building to run in parallel. This also brings the behavior of bytecode manifests in line with other manifests.

Reviewed By: IanChilds

Differential Revision: D74202392

fbshipit-source-id: 8dff5add6f8984b1d13c3f23bdcc69e55a99b46d