Add support for DocModule returned in the items of ForzenModule
Summary:
# This diff stack
Separate the rules at https://buck2.build/docs/prelude/globals/ into separate pages. Each page will have it's own page. With such change it will be more clean and it will be faster  to open the page.

# This diff

This diff add support return `DocModule` for the items of FrozenModule in `.documentation`.

In later diffs, we would use `namespace` to get the documentation. It would return `DocModule`.
Without this change, we would not get the docs for inner of namespace

Reviewed By: JakobDegen

Differential Revision: D73931900

fbshipit-source-id: 6f54cf7b489b0e870a088eabe0ff663c6d838057