erlang: change how include file information is stored
Summary:
This changes the storage in `BuildEnvironment` of includes from `dict[(app, path), dir]` to `dict[app, set[path]]`. In particular, this means that when we're merging environments from dependencies, we can mostly re-use the value of the field and don't need to deeply copy it and merge potentially massive dictionaries.
This new information is enough to do everything we were doing before with very similar efficiency (if not better) - in particular for private includes the new complexity is `O(apps)` where before it was `O(public_include_files)` (assuming lookup in a set is `O(1)`).

We still need the full mapping for `private_includes`, but most of the time that includes just the current application, except for tests - the overall storage for this is significantly smaller.

With this change total allocations drop by about 1.5GB or ~12%.

Reviewed By: TheGeorge

Differential Revision: D74246068

fbshipit-source-id: b48ddcb9d31293142fbc65cd5477cb9d292bdd22