Decouple remote dep file cache checking from action cache checking
Summary:
Previously, we just had one `cache_checker`, and if the action cache lookup failed, and remote dep file lookup was enabled, then we'd do a remote dep file lookup from within the action cache checker.

I want to change the order of the lookups that we do so that we check the local dep file cache before we do a remote dep file lookup, so this diff decouples the remote dep file lookup from the action cache lookup (but doesn't yet change the order). Now, we do two explicit calls in `run.rs`, first to the action cache and then to the remote dep file cache.

Reviewed By: blackm00n

Differential Revision: D74064170

fbshipit-source-id: 10c561ebe06dc87131a3eb113fa43cfdad78e051