Do local dep file cache lookup before a remote dep file cache lookup
Summary:
I think we should always use local dep files if we have them, as they are faster to access that remote dep files. This changes the order of the lookups.

I'm not sure if this needs a post in "Buck2 Dev", since it's not really an API change but it will affect some metrics. Tbh, I probably want to fix a bunch of metrics around this stuff anyway, so maybe worth a fuller post saying that I want to make this change along with some metrics changes. (E.g. right now I don't think we count identical dep file hits separately from actually dep file lookups in buck2 builds - though we do in buck2 actions).

Reviewed By: blackm00n

Differential Revision: D74064655

fbshipit-source-id: f6ad95f225c338c93f5b31ad3a70b57e90a06b75